# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenVizy API - Free music video generator
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TemplateStyle = typeof TemplateStyle[keyof typeof TemplateStyle];


export const TemplateStyle = {
  waveform: 'waveform',
  bars: 'bars',
  circular: 'circular',
  particles: 'particles',
  spectrum: 'spectrum',
  'album-float': 'album-float',
} as const;

export type TemplateSupportedFormatsItem = typeof TemplateSupportedFormatsItem[keyof typeof TemplateSupportedFormatsItem];


export const TemplateSupportedFormatsItem = {
  square: 'square',
  vertical: 'vertical',
  horizontal: 'horizontal',
} as const;

export interface Template {
  id: string;
  name: string;
  description: string;
  style: TemplateStyle;
  previewColor: string;
  supportedFormats: TemplateSupportedFormatsItem[];
  tags?: string[];
}

export type ProjectFormat = typeof ProjectFormat[keyof typeof ProjectFormat];


export const ProjectFormat = {
  square: 'square',
  vertical: 'vertical',
  horizontal: 'horizontal',
} as const;

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  pending: 'pending',
  processing: 'processing',
  ready: 'ready',
  failed: 'failed',
} as const;

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  artistName?: string | null;
  templateId: string;
  format: ProjectFormat;
  status: ProjectStatus;
  /** @nullable */
  audioFileName?: string | null;
  /** @nullable */
  artworkFileName?: string | null;
  /** @nullable */
  lyricsText?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
  /** @nullable */
  videoObjectPath?: string | null;
  createdAt: string;
}

export type ProjectInputFormat = typeof ProjectInputFormat[keyof typeof ProjectInputFormat];


export const ProjectInputFormat = {
  square: 'square',
  vertical: 'vertical',
  horizontal: 'horizontal',
} as const;

export interface ProjectInput {
  title: string;
  artistName?: string;
  templateId: string;
  format: ProjectInputFormat;
  lyricsText?: string;
  videoObjectPath?: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export interface ProjectStats {
  total: number;
  ready: number;
  processing: number;
  recentActivity: Project[];
}