# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OpenVizy API - Free music video generator
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Video project operations
  - name: templates
    description: Video template operations
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all available video templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all video projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new video project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a video project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /projects/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get summary stats for the dashboard
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Template:
      type: object
      required: [id, name, description, style, previewColor, supportedFormats]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        style:
          type: string
          enum: [waveform, bars, circular, particles, spectrum, album-float]
        previewColor:
          type: string
        supportedFormats:
          type: array
          items:
            type: string
            enum: [square, vertical, horizontal]
        tags:
          type: array
          items:
            type: string

    Project:
      type: object
      required: [id, title, templateId, format, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        artistName:
          type: ["string", "null"]
        templateId:
          type: string
        format:
          type: string
          enum: [square, vertical, horizontal]
        status:
          type: string
          enum: [pending, processing, ready, failed]
        audioFileName:
          type: ["string", "null"]
        artworkFileName:
          type: ["string", "null"]
        lyricsText:
          type: ["string", "null"]
        downloadUrl:
          type: ["string", "null"]
        videoObjectPath:
          type: ["string", "null"]
        createdAt:
          type: string

    ProjectInput:
      type: object
      required: [title, templateId, format]
      properties:
        title:
          type: string
        artistName:
          type: string
        templateId:
          type: string
        format:
          type: string
          enum: [square, vertical, horizontal]
        lyricsText:
          type: string
        videoObjectPath:
          type: string

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    ProjectStats:
      type: object
      required: [total, ready, processing, recentActivity]
      properties:
        total:
          type: integer
        ready:
          type: integer
        processing:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/Project"