# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenVizy API - Free music video generator
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from './project';

export interface ProjectStats {
  total: number;
  ready: number;
  processing: number;
  recentActivity: Project[];
}