# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenVizy API - Free music video generator
 * OpenAPI spec version: 0.1.0
 */
import type { TemplateStyle } from './templateStyle';
import type { TemplateSupportedFormatsItem } from './templateSupportedFormatsItem';

export interface Template {
  id: string;
  name: string;
  description: string;
  style: TemplateStyle;
  previewColor: string;
  supportedFormats: TemplateSupportedFormatsItem[];
  tags?: string[];
}