# lib/api-zod/src/generated/types/uploadUrlResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenVizy API - Free music video generator
 * OpenAPI spec version: 0.1.0
 */
import type { UploadUrlRequest } from './uploadUrlRequest';

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}